+++ qpsolvers/ecos_.py
@@ -24,6 +24,11 @@ from numpy.linalg import cholesky
 from scipy.sparse import csc_matrix
 
 
+def ecos_set_verbosity(verbose):
+    if verbose:
+        print("qpsolvers: `verbose` argument is not available for CVXPY yet")
+
+
 def ecos_solve_qp(P, q, G=None, h=None, A=None, b=None, initvals=None):
     """
     Solve a Quadratic Program defined as: