+++ qpsolvers/gurobi_.py
@@ -39,6 +39,11 @@ def get_nonzero_rows(M):
     return nonzero_rows
 
 
+def gurobi_set_verbosity(verbose):
+    if verbose:
+        print("qpsolvers: `verbose` argument is not available for CVXPY yet")
+
+
 def gurobi_solve_qp(P, q, G=None, h=None, A=None, b=None, initvals=None):
     """
     Solve a Quadratic Program defined as: