+++ qpsolvers/osqp_.py
@@ -32,6 +32,11 @@ def conversion_warning(M):
         "rather than as a numpy.ndarray" % (M, M)
 
 
+def osqp_set_verbosity(verbose):
+    if verbose:
+        print("qpsolvers: `verbose` argument is not available for CVXPY yet")
+
+
 def osqp_solve_qp(P, q, G=None, h=None, A=None, b=None, initvals=None):
     """
     Solve a Quadratic Program defined as: