+++ qpsolvers/qpoases_.py
@@ -31,6 +31,10 @@ options = Options()
 options.printLevel = PrintLevel.NONE
 
 
+def qpoases_set_verbosity(verbose):
+    options.printLevel = PrintLevel.MEDIUM if verbose else PrintLevel.NONE
+
+
 def qpoases_solve_qp(P, q, G=None, h=None, A=None, b=None, initvals=None,
                      max_wsr=1000):
     """